Connection feat

+++ ProjectFlightSchool/ProjectFlightSchool/Entry.cpp
@@ -52,14 +52,14 @@ bool Entry::Update( float deltaTime )
 	return true;
 }
 
-bool Entry::Initialize( UINT x, UINT y, UINT width, UINT height )
+bool Entry::Initialize( float x, float y, float width, float height )
 {
-	mUpperLeft.x	= x;
-	mUpperLeft.y	= y;
-	mLowerRight.x	= x + width;
-	mLowerRight.y	= y + height;
-	mWidth			= (float)width;
-	mHeight			= (float)height;
+	mUpperLeft.x	= (UINT)x;
+	mUpperLeft.y	= (UINT)y;
+	mLowerRight.x	= (UINT)x + width;
+	mLowerRight.y	= (UINT)y + height;
+	mWidth			= width;
+	mHeight			= height;
 	return true;
 }
