+++ ProjectFlightSchool/ProjectFlightSchool/Entry.h
@@ -28,7 +28,7 @@ class Entry
 		void			SetExitCooldown();
 		virtual void	Render() = 0;
 		virtual bool	Update( float deltaTime );
-		virtual bool	Initialize( UINT x, UINT y, UINT width, UINT height );
+		virtual bool	Initialize( float x, float y, float width, float height );
 		void			Release();
 						Entry();
 		virtual			~Entry();
